pyproject.toml
[project]
name = "ai-study-assistant"
version = "0.1.0"
description = "AI-powered study assistant built with Streamlit and Google Gemini."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.40.0",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
